feat: Implementando validacion de tokens y roles en las rutas

+++ src/routes/databaseConnection.routes.ts
@@ -1,10 +1,14 @@
 import { Router, Request, Response, NextFunction } from 'express';
 import DatabaseConnectionController from '../controllers/databaseConnection.controller.js';
+import { authenticate } from '../middlewares/auth.middleware.js';
+import { authorize } from '../middlewares/role.middleware.js';
 
 const DatabaseConnectionRouter = Router();
 
 DatabaseConnectionRouter.get(
   '/',
+  authenticate,
+  authorize(['Admin', 'Client', 'User']),
   (req: Request, res: Response, next: NextFunction) => {
     DatabaseConnectionController.getAllConnections(req, res, next);
   }
@@ -12,6 +16,8 @@ DatabaseConnectionRouter.get(
 
 DatabaseConnectionRouter.get(
   '/:id',
+  authenticate,
+  authorize(['Admin', 'Client', 'User']),
   (req: Request, res: Response, next: NextFunction) => {
     DatabaseConnectionController.getConnectionById(req, res, next);
   }
@@ -19,6 +25,8 @@ DatabaseConnectionRouter.get(
 
 DatabaseConnectionRouter.post(
   '/',
+  authenticate,
+  authorize(['Admin', 'Client', 'User']),
   (req: Request, res: Response, next: NextFunction) => {
     DatabaseConnectionController.createConnection(req, res, next);
   }
@@ -26,6 +34,8 @@ DatabaseConnectionRouter.post(
 
 DatabaseConnectionRouter.put(
   '/:id',
+  authenticate,
+  authorize(['Admin']),
   (req: Request, res: Response, next: NextFunction) => {
     DatabaseConnectionController.updateConnection(req, res, next);
   }
@@ -33,6 +43,8 @@ DatabaseConnectionRouter.put(
 
 DatabaseConnectionRouter.delete(
   '/:id',
+  authenticate,
+  authorize(['Admin']),
   (req: Request, res: Response, next: NextFunction) => {
     DatabaseConnectionController.deleteConnection(req, res, next);
   }
