+++ src/routes/queryTemplate.routes.ts
@@ -1,10 +1,14 @@
 import { Router, Request, Response, NextFunction } from 'express';
 import QueryTemplateController from '../controllers/queryTemplate.controller.js';
+import { authenticate } from '../middlewares/auth.middleware.js';
+import { authorize } from '../middlewares/role.middleware.js';
 
 const QueryTemplateRouter = Router();
 
 QueryTemplateRouter.get(
   '/',
+  authenticate,
+  authorize(['Admin', 'Client', 'User']),
   (req: Request, res: Response, next: NextFunction) => {
     QueryTemplateController.getAllQueryTemplates(req, res, next);
   }
@@ -12,6 +16,8 @@ QueryTemplateRouter.get(
 
 QueryTemplateRouter.get(
   '/:id',
+  authenticate,
+  authorize(['Admin', 'Client', 'User']),
   (req: Request, res: Response, next: NextFunction) => {
     QueryTemplateController.getQueryTemplateById(req, res, next);
   }
@@ -19,6 +25,8 @@ QueryTemplateRouter.get(
 
 QueryTemplateRouter.post(
   '/',
+  authenticate,
+  authorize(['Admin', 'Client', 'User']),
   (req: Request, res: Response, next: NextFunction) => {
     QueryTemplateController.createQueryTemplate(req, res, next);
   }
@@ -26,6 +34,8 @@ QueryTemplateRouter.post(
 
 QueryTemplateRouter.put(
   '/:id',
+  authenticate,
+  authorize(['Admin', 'Client', 'User']),
   (req: Request, res: Response, next: NextFunction) => {
     QueryTemplateController.updateQueryTemplate(req, res, next);
   }
@@ -33,6 +43,8 @@ QueryTemplateRouter.put(
 
 QueryTemplateRouter.delete(
   '/:id',
+  authenticate,
+  authorize(['Admin', 'Client', 'User']),
   (req: Request, res: Response, next: NextFunction) => {
     QueryTemplateController.deleteQueryTemplate(req, res, next);
   }
