+++ src/routes/scheduledQuery.routes.ts
@@ -1,10 +1,14 @@
 import { Router, Request, Response, NextFunction } from 'express';
 import ScheduledQueryController from '../controllers/scheduledQuery.controller.js';
+import { authenticate } from '../middlewares/auth.middleware.js';
+import { authorize } from '../middlewares/role.middleware.js';
 
 const ScheduledQueryRouter = Router();
 
 ScheduledQueryRouter.get(
   '/',
+  authenticate,
+  authorize(['Admin', 'Client', 'User']),
   (req: Request, res: Response, next: NextFunction) => {
     ScheduledQueryController.getAllScheduledQueries(req, res, next);
   }
@@ -12,6 +16,8 @@ ScheduledQueryRouter.get(
 
 ScheduledQueryRouter.get(
   '/:id',
+  authenticate,
+  authorize(['Admin', 'Client', 'User']),
   (req: Request, res: Response, next: NextFunction) => {
     ScheduledQueryController.getScheduledQueryById(req, res, next);
   }
@@ -19,6 +25,8 @@ ScheduledQueryRouter.get(
 
 ScheduledQueryRouter.post(
   '/',
+  authenticate,
+  authorize(['Admin', 'Client', 'User']),
   (req: Request, res: Response, next: NextFunction) => {
     ScheduledQueryController.createScheduledQuery(req, res, next);
   }
@@ -26,6 +34,8 @@ ScheduledQueryRouter.post(
 
 ScheduledQueryRouter.put(
   '/:id',
+  authenticate,
+  authorize(['Admin', 'Client', 'User']),
   (req: Request, res: Response, next: NextFunction) => {
     ScheduledQueryController.updateScheduledQuery(req, res, next);
   }
@@ -33,6 +43,8 @@ ScheduledQueryRouter.put(
 
 ScheduledQueryRouter.delete(
   '/:id',
+  authenticate,
+  authorize(['Admin', 'Client', 'User']),
   (req: Request, res: Response, next: NextFunction) => {
     ScheduledQueryController.deleteScheduledQuery(req, res, next);
   }
